Added generic lifetimes to VectorType trait
The generic type of this trait will be useful, as it will
reduce a lot of boiler-plate code. The intended use of the
generic type is to be the inner type of a vector type.

The trait constraint was updated to guarentee a borrowed item when
iterating over the vector. This property will be directly used
for future traits.